+++ bfabric_app_runner/src/bfabric_app_runner/specs/inputs/bfabric_annotation_spec.py
@@ -2,9 +2,8 @@ from __future__ import annotations
 
 from typing import Literal, TYPE_CHECKING, Annotated
 
-from pydantic import BaseModel, Field
-
 from bfabric_app_runner.specs.common_types import RelativeFilePath  # noqa: TC001
+from pydantic import BaseModel, Field
 
 if TYPE_CHECKING:
     from bfabric import Bfabric
@@ -22,6 +21,8 @@ class BfabricAnnotationResourceSampleSpec(_AnnotationSpec):
     annotation: Literal["resource_sample"] = "resource_sample"
     separator: str
     resource_ids: list[int]
+    format: Literal["csv"] = "csv"
 
 
-BfabricAnnotationSpec = Annotated[BfabricAnnotationResourceSampleSpec, Field(discriminator="annotation")]
+BfabricAnnotationSpec = BfabricAnnotationResourceSampleSpec
+BfabricAnnotationSpecField = Annotated[BfabricAnnotationSpec, Field(discriminator="annotation")]