+++ bfabric_scripts/src/bfabric_scripts/cli/dataset/show.py
@@ -45,8 +45,7 @@ def _print_yaml(dataframe: pl.DataFrame) -> None:
 def cmd_dataset_show(dataset_id: int, format: OutputFormat = OutputFormat.TABLE, *, client: Bfabric) -> None:
     """Show a dataset in the console."""
     dataset = Dataset.find(id=dataset_id, client=client)
-    # TODO fix
-    types = dataset.types
+    types = dataset.column_types
     dataframe = dataset.to_polars()
 
     if format == OutputFormat.TABLE: